+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* fetchHomeFeed() has skipMutex=true
 
 ### v0.29.2
 * Add `Config.lookbackForUpdatesMinutes` param and scan backwards from the maxUpdatedAt that many more minutes to catch changes